Renamed controllers to grpc services
